Rearrange unit tests
